# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job Tracker Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/subscription:
    get:
      operationId: getSubscription
      tags: [payments]
      summary: Get current user subscription
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    post:
      operationId: updateSubscription
      tags: [payments]
      summary: Upgrade or downgrade plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, planId, planName, status, currentPeriodEnd]
      properties:
        id:
          type: integer
        planId:
          type: string
        planName:
          type: string
        status:
          type: string
        currentPeriodEnd:
          type: string
        cancelAtPeriodEnd:
          type: boolean
    SubscriptionInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
...